+++ src/main/java/data/SerialHandler.java
@@ -4,17 +4,14 @@
 package data;
 
 import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
 import java.io.Serializable;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 
 /**
- *
+ * Class to write/read StockItem list serialized
  * @author blegendre
  */
 public class SerialHandler {